Refactored core systems to improve modularity and maintainability:
- Updated `Controller` and `GameSessionData` to Singleton pattern.
- Enhanced screen management with stack-based logic and new screens (`Market`, `Help`).
- Overhauled weapon system with specific weapon types (`HandGun`, `Shotgun`, `Rifle`, `Sniper`) and added firing logic, damage, and light effects.
- Improved `Bullet`, `Character`, `Enemy`, and `Player` classes with new properties and methods for dynamic gameplay.

Simplified and cleaned up codebase:
- Removed legacy code (`Gun`, `SwitchScreenCommand`) and unused methods.
- Refactored `Bar`, `Entity`, and `State` classes for better consistency.
- Updated `CMakeLists.txt` to include new assets and source files.

Added new features:
- Introduced `Help` screen with scrolling and user guidance.
- Added `Market` screen for weapon purchases and inventory management.
- Improved button rendering and text updates using shared font.

Miscellaneous:
- Fixed typos, improved naming conventions, and enhanced code readability.
- Adjusted collision handling with new `bodyRadius` property in multiple classes.

+++ include/HealthBar.h
@@ -4,6 +4,6 @@
 
 class HealthBar : public  Bar{
 public:
-    HealthBar(float width, float height, int maxHealth);
+    HealthBar(float width, float height, float maxHealth);
 	void updateColor() override;
 };
